Splitting Pipeline.scala into three separate files.
