+++ src/main/scala/pipelines/Pipeline.scala
@@ -1,7 +1,7 @@
 package pipelines
 
 import org.apache.spark.rdd.RDD
-import pipelines.Pipelines.PipelineNode
+
 
 import scala.reflect.ClassTag
 
@@ -159,131 +159,3 @@ object Pipeline {
    */
   def apply[A : ClassTag, L]() = new Pipeline[A, A, A, L](new IdentityTransformer[A], new IdentityTransformer[A])
 }
-
-/**
- * Creates a chain of transformers. (This is only used internally).
- * @param a First transformer.
- * @param b Second transformer.
- * @tparam A Transformer chain input type.
- * @tparam B Intermediate type.
- * @tparam C Output type.
- */
-private class TransformerChain[A : ClassTag, B, C : ClassTag](
-    val a: Transformer[A, B],
-    val b: Transformer[B, C]) extends Transformer[A, C] {
-
-  /**
-   * Transform an input.
-   * @param in Input.
-   * @return Transformed input.
-   */
-  override def transform(in: A): C = b.transform(a.transform(in))
-
-  /**
-   * Transform an RDD of Inputs.
-   * @param in Input RDD.
-   * @return RDD of transformed inputs.
-   */
-  override def transform(in: RDD[A]): RDD[C] = b.transform(a.transform(in))
-}
-
-/**
- * Special kind of transformer that does nothing.
- */
-class IdentityTransformer[T : ClassTag] extends Transformer[T, T] {
-  override def transform(in: T): T = in
-  override def transform(in: RDD[T]): RDD[T] = in
-}
-
-/**
- * A transformer is a deterministic object that takes data of type A and turns it into type B.
- * It differs from a function in that it can operate on RDDs.
- *
- * @tparam A Input type.
- * @tparam B Output type.
- */
-abstract class Transformer[A : ClassTag, B : ClassTag]
-    extends PipelineNode[RDD[A], RDD[B]]
-    with Serializable
-    with PipelineStage[A, B] {
-
-  /**
-   * Takes an A and returns a B.
-   *
-   * @param in Input.
-   * @return Output.
-   */
-  def transform(in: A): B
-
-  /**
-   * Takes an RDD[A] and returns an RDD[B].
-   *
-   * @param in Collection of A's.
-   * @return Collection of B's.
-   */
-  def transform(in: RDD[A]): RDD[B] = in.map(transform)
-
-  /**
-   * Syntactic sugar for transform().
-   *
-   * @param in Collection of A's.
-   * @return Collection of B's.
-   */
-  def apply(in: RDD[A]): RDD[B] = transform(in)
-
-  /**
-   * Append a transformer onto another transformer, forming a transformer chain.
-   * This is the logical equivalent of function composition.
-   *
-   * @param next Second transformer.
-   * @tparam C Output type of second transformer.
-   * @return A new transformer chain.
-   */
-  def andThen[C : ClassTag](next: Transformer[B, C]): Transformer[A, C] = new TransformerChain[A, B, C](this, next)
-}
-
-/**
- * An Unsupervised estimator. An estimator can emit a transformer via a call to its fit method on unlabeled data.
- *
- * @tparam I Input type.
- * @tparam O Output type.
- */
-abstract class UnsupervisedEstimator[I, O] extends Serializable with PipelineStage[I, O] {
-  def fit(in: RDD[I]): Transformer[I, O]
-}
-
-/**
- * A Supervised estimator. A supervised estimator emits a transformer via a call to its fit method on labeled data.
- *
- * @tparam I Input type.
- * @tparam O Output type.
- * @tparam L Label type.
- */
-abstract class SupervisedEstimator[I, O, L] extends Serializable with PipelineStage[I, O] {
-  def fit(data: RDD[I], labels: RDD[L]): Transformer[I, O]
-}
-
-/**
- * An unsupervised estimator that carries its data with it.
- * @param e Estimator.
- * @param data Data.
- * @tparam I Input type.
- * @tparam O Output type.
- */
-private case class UnsupervisedEstimatorWithData[I, O](
-    e: UnsupervisedEstimator[I, O],
-    data: RDD[I]) extends PipelineStage[I, O]
-
-/**
- * A supervised estimator that carries its data with it.
- * @param e Estimator.
- * @param data Data.
- * @param labels Labels.
- * @tparam I Input type.
- * @tparam O Output type.
- * @tparam L Label type.
- */
-private case class SupervisedEstimatorWithData[I, O, L](
-    e: SupervisedEstimator[I, O, L],
-    data: RDD[I],
-    labels: RDD[L]) extends PipelineStage[I, O]
